refactor: rename pages
